change some strategic casts to (*long) from (*int).

+++ sys/vm/vm_extern.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: vm_extern.h,v 1.9 1994/10/20 04:27:30 cgd Exp $	*/
+/*	$NetBSD: vm_extern.h,v 1.10 1994/10/29 07:35:07 cgd Exp $	*/
 
 /*-
  * Copyright (c) 1992, 1993
@@ -91,8 +91,8 @@ void		 swapout_threads __P((void));
 int		 swfree __P((struct proc *, int));
 void		 swstrategy __P((struct buf *));
 void		 thread_block __P((void));
-void		 thread_sleep __P((int, simple_lock_t, boolean_t));
-void		 thread_wakeup __P((int));
+void		 thread_sleep __P((long, simple_lock_t, boolean_t));
+void		 thread_wakeup __P((long));
 int		 useracc __P((caddr_t, int, int));
 int		 vm_allocate __P((vm_map_t,
 		    vm_offset_t *, vm_size_t, boolean_t));
